requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.12.0
matplotlib>=3.5.0
scipy>=1.7.0
tqdm>=4.62.0
pyyaml>=6.0
typer>=0.9.0
rich>=13.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
